feat: EmptyPlaceHolder 컴포넌트 생성 close#342
